prerequisite items for new gameplay
