+++ Assets/Scripts/Game/Entities/EntityDataManager.cs
@@ -13,14 +13,6 @@ namespace Celeritas.Game.Entities
 	/// </summary>
 	public class EntityDataManager : Singleton<EntityDataManager>
 	{
-		private const float CHUNK_SIZE = 100;
-
-		private const int DISABLE_CHUNK_DIST = 3;
-
-		private const int UNLOAD_CHUNK_DIST = 6;
-
-		private const float UPDATE_FREQ = 2;
-
 		private readonly List<ShipData> ships = new List<ShipData>();
 		private readonly List<WeaponData> weapons = new List<WeaponData>();
 		private readonly List<ModuleData> modules = new List<ModuleData>();
@@ -88,103 +80,18 @@ namespace Celeritas.Game.Entities
 		/// </summary>
 		public IReadOnlyList<ShipData> PlayerShips { get => playerShips.AsReadOnly(); }
 
-		/// <summary>
-		/// Has the game loaded its assets.
-		/// </summary>
-		public bool Loaded { get; private set; } = false;
-
-		/// <summary>
-		/// The chunk manager for the game.
-		/// </summary>
-		public static ChunkManager ChunkManager { get; private set; }
+		public bool Loaded { get; private set; }
 
 		public static event Action OnLoadedAssets;
 
 		public static event Action<Entity> OnCreatedEntity;
 
-		public static event Action<Chunk> OnCreatedChunk;
-
 		private static Dictionary<EntityData, ObjectPool<Entity>> entites = new Dictionary<EntityData, ObjectPool<Entity>>();
-
-		private float lastUpdate;
-
-		private new Camera camera;
 
 		protected override void Awake()
 		{
 			base.Awake();
-			camera = Camera.main;
-			ChunkManager = new ChunkManager(new Vector2(CHUNK_SIZE, CHUNK_SIZE));
 			LoadAssets();
-		}
-
-		private void FixedUpdate()
-		{
-			if (!Instance.Loaded || Time.unscaledTime < lastUpdate + (1f / UPDATE_FREQ))
-				return;
-
-			lastUpdate = Time.unscaledTime;
-			UpdateChunks();
-		}
-
-		private Color green = new Color(0, 1, 0, 0.1f);
-		private Color yellow = new Color(1, 0.92f, 0.016f, 0.1f);
-
-		private void OnDrawGizmosSelected()
-		{
-			if (ChunkManager == null)
-				return;
-
-			foreach (var chunk in ChunkManager.Chunks)
-			{
-				if (chunk.Active)
-					Gizmos.color = green;
-				else
-					Gizmos.color = yellow;
-
-				Gizmos.DrawCube(chunk.Center, new Vector3(chunk.Size.x, chunk.Size.y, 1));
-			}
-		}
-
-		private void UpdateChunks()
-		{
-			var middle = ChunkManager.GetChunkIndex(camera.transform.position);
-
-			for (int x = 0; x < UNLOAD_CHUNK_DIST * 2; x++)
-			{
-				for (int y = 0; y < UNLOAD_CHUNK_DIST * 2; y++)
-				{
-					var index = middle + new Vector2Int(x - UNLOAD_CHUNK_DIST, y - UNLOAD_CHUNK_DIST);
-
-					if (ChunkManager.GetManhattenDistance(middle, index) >= UNLOAD_CHUNK_DIST)
-						continue;
-
-					if (ChunkManager.TryGetChunk(index, out var chunk))
-					{
-						chunk.ChunkSetActive(ChunkManager.GetManhattenDistance(middle, index) < DISABLE_CHUNK_DIST);
-					}
-					else
-					{
-						chunk = ChunkManager.CreateChunk(index);
-						chunk.ChunkSetActive(ChunkManager.GetManhattenDistance(middle, index) < DISABLE_CHUNK_DIST);
-						OnCreatedChunk?.Invoke(chunk);
-					}
-				}
-			}
-
-			var toRemove = new HashSet<Vector2Int>();
-			foreach (var chunk in ChunkManager.Keys)
-			{
-				if (ChunkManager.GetManhattenDistance(middle, chunk) >= UNLOAD_CHUNK_DIST)
-				{
-					toRemove.Add(chunk);
-				}
-			}
-
-			foreach (var chunk in toRemove)
-			{
-				ChunkManager.UnloadChunk(chunk);
-			}
 		}
 
 		/// <summary>
@@ -215,13 +122,13 @@ namespace Celeritas.Game.Entities
 
 			if (data.UseChunking)
 			{
-				if (ChunkManager.TryGetChunk(position, out var chunk))
+				if (Chunks.ChunkManager.TryGetChunk(position, out var chunk))
 				{
 					chunk.AddEntity(entity);
 				}
 				else
 				{
-					UnityEngine.Debug.LogError($"Tried to spawn \"{data.Title}\" in a chunk which does not exist ({ChunkManager.GetChunkIndex(position).x}, {ChunkManager.GetChunkIndex(position).y}). " +
+					UnityEngine.Debug.LogError($"Tried to spawn \"{data.Title}\" in a chunk which does not exist ({Chunks.ChunkManager.GetChunkIndex(position).x}, {Chunks.ChunkManager.GetChunkIndex(position).y}). " +
 						$"Check spawn logic or disable \"useChunking\" for this object.", entity.gameObject);
 				}
 			}
@@ -334,7 +241,6 @@ namespace Celeritas.Game.Entities
 			UnityEngine.Debug.Log($"load took: {watch.ElapsedMilliseconds}ms");
 
 			Loaded = true;
-			UpdateChunks();
 			OnLoadedAssets?.Invoke();
 		}
