+++ Assets/Scripts/Game/Events.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: fbb6e7f2fe5aa2f43ae5fe83bc489c21
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
