+++ Assets/Scripts/Game/Events/EnemyEventData.cs
@@ -1,0 +1,12 @@
+using System.Collections;
+using System.Collections.Generic;
+using UnityEngine;
+
+namespace Celeritas.Game.Events
+{
+	[CreateAssetMenu(fileName = "New Event", menuName = "Celeritas/Events/New Enemy Event", order = 10)]
+	public class EnemyEventData : EventData
+	{
+
+	}
+}
