+++ Assets/Scripts/Game/Events/EnemyEventData.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 2632016ce39037c4eb07668fe705fb9c
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
