+++ Assets/Scripts/Game/Events/EventData.cs
@@ -1,0 +1,15 @@
+using System.Collections;
+using System.Collections.Generic;
+using UnityEngine;
+
+namespace Celeritas.Game.Events
+{
+	public abstract class EventData : ScriptableObject
+	{
+		[SerializeField]
+		private bool showOnMap;
+
+		[SerializeField]
+		private bool showArrow;
+	}
+}
