+++ Assets/Scripts/Game/Events/EventData.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 4db6201628f752148971b3fd2430c3a5
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
