+++ Assets/Scripts/Game/Events/GameEvent.cs
@@ -1,0 +1,16 @@
+using System.Collections;
+using System.Collections.Generic;
+using UnityEngine;
+
+namespace Celeritas.Game.Events
+{
+	public class GameEvent
+	{
+		public EventData EventData { get; private set; }
+
+		public virtual void Initalize(EventData data)
+		{
+			EventData = data;
+		}
+	}
+}
