+++ Assets/Scripts/Game/Events/GameEvent.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: f9895a59ae9bd72418f38361fb2ed25e
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
